Make `library-build` into a fixture
