+++ fixtures/library-build/.eslintignore
