+++ fixtures/library-build/.gitignore
