+++ fixtures/library-build/src/library.js
