+++ fixtures/library-build/src/render.js
