+++ fixtures/library-build/src/styles.less.d.ts
